+++ packages/typespec-rust/src/tcgcadapter/adapter.ts
@@ -1170,8 +1170,11 @@ export class Adapter {
         adaptedParam = this.adaptMethodParameter(opParam);
       }
 
-      if (adaptedParam.kind === 'headerScalar' || adaptedParam.kind === 'queryScalar') {
-        paramsMap.set(param, adaptedParam);
+      switch (adaptedParam.kind) {
+        case 'headerScalar':
+        case 'queryScalar':
+          paramsMap.set(param, adaptedParam);
+          break;
       }
 
       adaptedParam.docs = this.adaptDocs(param.summary, param.doc);
@@ -1664,13 +1667,11 @@ export class Adapter {
       case 'path': {
         paramType = this.typeToWireType(paramType);
         let style: rust.ParameterStyle = 'simple';
-        {
-          const tspStyleString = (param.style as string);
-          if (!['simple', 'path', 'label', 'matrix'].includes(tspStyleString)) {
-            throw new AdapterError('InternalError', `unsupported style ${tspStyleString} for parameter ${param.serializedName}`, param.__raw?.node);
-          } else {
-            style = tspStyleString as rust.ParameterStyle;
-          }
+        const tspStyleString = (param.style as string);
+        if (!['simple', 'path', 'label', 'matrix'].includes(tspStyleString)) {
+          throw new AdapterError('InternalError', `unsupported style ${tspStyleString} for parameter ${param.serializedName}`, param.__raw?.node);
+        } else {
+          style = tspStyleString as rust.ParameterStyle;
         }
 
         if (isRefSlice(paramType)) {